Modify curl private header to use siz_t for buffers.

+++ sdk/core/azure-core/src/http/curl/curl_session_private.hpp
@@ -146,7 +146,7 @@ namespace Azure { namespace Core { namespace Http {
        * @param bufferSize Indicates the size of the buffer.
        * @return Returns the index of the last parsed position from buffer.
        */
-      int64_t BuildStatusCode(uint8_t const* const buffer, int64_t const bufferSize);
+      size_t BuildStatusCode(uint8_t const* const buffer, size_t const bufferSize);
 
       /**
        * @brief This method is invoked by the Parsing process if the internal state is set to
@@ -159,7 +159,7 @@ namespace Azure { namespace Core { namespace Http {
        * value is smaller than the body size, means there is part of the body response in the
        * buffer.
        */
-      int64_t BuildHeader(uint8_t const* const buffer, int64_t const bufferSize);
+      size_t BuildHeader(uint8_t const* const buffer, size_t const bufferSize);
 
     public:
       /**
@@ -183,7 +183,7 @@ namespace Azure { namespace Core { namespace Http {
        * RawResponse is completed and that the rest of the buffer contains part of the response
        * body.
        */
-      int64_t Parse(uint8_t const* const buffer, int64_t const bufferSize);
+      size_t Parse(uint8_t const* const buffer, size_t const bufferSize);
 
       /**
        * @brief Indicates when the parser has completed parsing and building the HTTP RawResponse.
@@ -239,7 +239,7 @@ namespace Azure { namespace Core { namespace Http {
      * decide how much data to take from the inner buffer before pulling more data from network.
      *
      */
-    int64_t m_bodyStartInBuffer = -1;
+    size_t m_bodyStartInBuffer = -1;
 
     /**
      * @brief Control field to handle the number of bytes containing relevant data within the
@@ -247,7 +247,7 @@ namespace Azure { namespace Core { namespace Http {
      * from wire into it, it can be holding less then N bytes.
      *
      */
-    int64_t m_innerBufferSize = _detail::DefaultLibcurlReaderSize;
+    size_t m_innerBufferSize = _detail::DefaultLibcurlReaderSize;
 
     bool m_isChunkedResponseType = false;
